requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.10
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-mistralai==0.2.3
langchain-text-splitters==0.3.2
langsmith==0.1.147
mando==0.7.1
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.1.3
openai==1.56.0
orjson==3.10.12
packaging==24.2
prompt_toolkit==3.0.48
propcache==0.2.1
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
radon==6.0.1
redis==5.2.0
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.0.0
tokenizers==0.21.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
